container/dags/requirements.txt
apache-airflow[cncf.kubernetes]
boto3
pandas
pathlib2
numpy
scipy
psycopg2
sqlalchemy
pyparsing
urllib3